requirements.txt
streamlit
pdfplumber
python-docx
beautifulsoup4
requests
lxml
spacy
sentence-transformers
faiss-cpu
langchain
numpy
scikit-learn
unstructured